Splitting configuration files (one for each dashboard)
