+++ app/javascript/gobierto_dashboards/webapp/containers/subsidies/Aside.vue
@@ -42,7 +42,7 @@
 import { BlockHeader, Checkbox, Dropdown } from "lib/vue-components";
 import DownloadButton from "../../components/DownloadButton.vue";
 import { EventBus } from "../../mixins/event_bus";
-import { subsidiesFiltersConfig } from "../../lib/config.js";
+import { subsidiesFiltersConfig } from "../../lib/config/subsidies.js";
 
 export default {
   name: 'Aside',
