+++ app/javascript/gobierto_dashboards/webapp/lib/config/contracts.js
@@ -1,3 +1,4 @@
+// table columns
 export const contractsColumns = [
   {field: 'assignee', translation: I18n.t('gobierto_dashboards.dashboards.contracts.assignee'), format: null, cssClass: ''},
   {field: 'title', translation: I18n.t('gobierto_dashboards.dashboards.contracts.contractor'), format: 'truncated', cssClass: 'largest-width-td'},
@@ -23,18 +24,7 @@ export const assigneesShowColumns = [
   {field: 'start_date', translation: I18n.t('gobierto_dashboards.dashboards.contracts.date'), format: null, cssClass: ''},
 ];
 
-export const subsidiesColumns = [
-  {field: 'beneficiary_name', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.beneficiary'), format: null, cssClass: ''},
-  {field: 'amount', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.amount'), format: 'currency', cssClass: 'right'},
-  {field: 'grant_date', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.date'), format: null, cssClass: ''},
-];
-
-export const grantedColumns = [
-  {field: 'name', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.beneficiary'), format: null, cssClass: ''},
-  {field: 'count', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.subsidies'), format: null, cssClass: 'right'},
-  {field: 'sum', translation: I18n.t('gobierto_dashboards.dashboards.subsidies.amount'), format: 'currency', cssClass: 'right'},
-];
-
+// filters config
 export const contractsFiltersConfig = [
   {
     id: 'dates',
@@ -51,21 +41,6 @@ export const contractsFiltersConfig = [
   {
     id: 'process_types',
     title: I18n.t('gobierto_dashboards.dashboards.contracts.process_type'),
-    isToggle: true,
-    options: []
-  }
-]
-
-export const subsidiesFiltersConfig = [
-  {
-    id: 'dates',
-    title: I18n.t('gobierto_dashboards.dashboards.subsidies.dates'),
-    isToggle: true,
-    options: []
-  },
-  {
-    id: 'categories',
-    title: I18n.t('gobierto_dashboards.dashboards.subsidies.category'),
     isToggle: true,
     options: []
   }
